+++ cell/cell.go
@@ -23,8 +23,14 @@ type Option interface {
 
 // Options stores the provided options.
 type Options struct {
-	FgColor Color
-	BgColor Color
+	FgColor       Color
+	BgColor       Color
+	Bold          bool
+	Italic        bool
+	Underline     bool
+	Strikethrough bool
+	Inverse       bool
+	Blink         bool
 }
 
 // Set allows existing options to be passed as an option.
@@ -60,5 +66,47 @@ func FgColor(color Color) Option {
 func BgColor(color Color) Option {
 	return option(func(co *Options) {
 		co.BgColor = color
+	})
+}
+
+// Bold makes cell's text bold.
+func Bold() Option {
+	return option(func(co *Options) {
+		co.Bold = true
+	})
+}
+
+// Italic makes cell's text italic. Only works when using the tcell backend.
+func Italic() Option {
+	return option(func(co *Options) {
+		co.Italic = true
+	})
+}
+
+// Underline makes cell's text underlined.
+func Underline() Option {
+	return option(func(co *Options) {
+		co.Underline = true
+	})
+}
+
+// Strikethrough strikes through the cell's text. Only works when using the tcell backend.
+func Strikethrough() Option {
+	return option(func(co *Options) {
+		co.Strikethrough = true
+	})
+}
+
+// Inverse inverts the colors of the cell's text.
+func Inverse() Option {
+	return option(func(co *Options) {
+		co.Inverse = true
+	})
+}
+
+// Blink makes the cell's text blink. Only works when using the tcell backend.
+func Blink() Option {
+	return option(func(co *Options) {
+		co.Blink = true
 	})
 }