+++ termdashdemo/termdashdemo.go
@@ -20,6 +20,7 @@ import (
 	"context"
 	"flag"
 	"fmt"
+	"image"
 	"log"
 	"math"
 	"math/rand"
@@ -69,9 +70,9 @@ type widgets struct {
 }
 
 // newWidgets creates all widgets used by this demo.
-func newWidgets(ctx context.Context, c *container.Container) (*widgets, error) {
+func newWidgets(ctx context.Context, t terminalapi.Terminal, c *container.Container) (*widgets, error) {
 	updateText := make(chan string)
-	sd, err := newSegmentDisplay(ctx, updateText)
+	sd, err := newSegmentDisplay(ctx, t, updateText)
 	if err != nil {
 		return nil, err
 	}
@@ -477,8 +478,8 @@ const (
 
 func main() {
 	terminalPtr := flag.String("terminal",
-		"termbox",
-		"The terminal implementation to use. Available implementations are 'termbox' and 'tcell' (default = termbox).")
+		"tcell",
+		"The terminal implementation to use. Available implementations are 'termbox' and 'tcell' (default = tcell).")
 	flag.Parse()
 
 	var t terminalapi.Terminal
@@ -504,7 +505,7 @@ func main() {
 	}
 
 	ctx, cancel := context.WithCancel(context.Background())
-	w, err := newWidgets(ctx, c)
+	w, err := newWidgets(ctx, t, c)
 	if err != nil {
 		panic(err)
 	}
@@ -569,7 +570,7 @@ func textState(text string, capacity, step int) []rune {
 // SegmentDisplay.
 func newTextInput(updateText chan<- string) (*textinput.TextInput, error) {
 	input, err := textinput.New(
-		textinput.Label("Change text to: ", cell.FgColor(cell.ColorBlue)),
+		textinput.Label("Change text to: ", cell.FgColor(cell.ColorNumber(33))),
 		textinput.MaxWidthCells(20),
 		textinput.PlaceHolder("enter any text"),
 		textinput.OnSubmit(func(text string) error {
@@ -586,17 +587,17 @@ func newTextInput(updateText chan<- string) (*textinput.TextInput, error) {
 
 // newSegmentDisplay creates a new SegmentDisplay that initially shows the
 // Termdash name. Shows any text that is sent over the channel.
-func newSegmentDisplay(ctx context.Context, updateText <-chan string) (*segmentdisplay.SegmentDisplay, error) {
+func newSegmentDisplay(ctx context.Context, t terminalapi.Terminal, updateText <-chan string) (*segmentdisplay.SegmentDisplay, error) {
 	sd, err := segmentdisplay.New()
 	if err != nil {
 		return nil, err
 	}
 
 	colors := []cell.Color{
-		cell.ColorBlue,
+		cell.ColorNumber(33),
 		cell.ColorRed,
 		cell.ColorYellow,
-		cell.ColorBlue,
+		cell.ColorNumber(33),
 		cell.ColorGreen,
 		cell.ColorRed,
 		cell.ColorGreen,
@@ -609,12 +610,33 @@ func newSegmentDisplay(ctx context.Context, updateText <-chan string) (*segmentd
 	go func() {
 		ticker := time.NewTicker(500 * time.Millisecond)
 		defer ticker.Stop()
+
+		capacity := 0
+		termSize := t.Size()
 		for {
 			select {
 			case <-ticker.C:
-				state := textState(text, sd.Capacity(), step)
+				if capacity == 0 {
+					// The segment display only knows its capacity after both
+					// text size and terminal size are known.
+					capacity = sd.Capacity()
+				}
+				if t.Size().Eq(image.ZP) || !t.Size().Eq(termSize) {
+					// Update the capacity initially the first time the
+					// terminal reports a non-zero size and then every time the
+					// terminal resizes.
+					//
+					// This is better than updating the capacity on every
+					// iteration since that leads to edge cases - segment
+					// display capacity depends on the length of text and here
+					// we are trying to adjust the text length to the capacity.
+					termSize = t.Size()
+					capacity = sd.Capacity()
+				}
+
+				state := textState(text, capacity, step)
 				var chunks []*segmentdisplay.TextChunk
-				for i := 0; i < sd.Capacity(); i++ {
+				for i := 0; i < capacity; i++ {
 					if i >= len(state) {
 						break
 					}
@@ -854,7 +876,7 @@ func newSines(ctx context.Context) (left, right *button.Button, lc *linechart.Li
 	go periodic(ctx, redrawInterval/3, func() error {
 		step1 = (step1 + 1) % len(inputs)
 		if err := lc.Series("first", rotateFloats(inputs, step1),
-			linechart.SeriesCellOpts(cell.FgColor(cell.ColorBlue)),
+			linechart.SeriesCellOpts(cell.FgColor(cell.ColorNumber(33))),
 		); err != nil {
 			return err
 		}