+++ widgets/textinput/textinputdemo/textinputdemo.go
@@ -26,7 +26,7 @@ import (
 	"github.com/mum4k/termdash/container/grid"
 	"github.com/mum4k/termdash/keyboard"
 	"github.com/mum4k/termdash/linestyle"
-	"github.com/mum4k/termdash/terminal/termbox"
+	"github.com/mum4k/termdash/terminal/tcell"
 	"github.com/mum4k/termdash/widgets/button"
 	"github.com/mum4k/termdash/widgets/segmentdisplay"
 	"github.com/mum4k/termdash/widgets/textinput"
@@ -60,10 +60,10 @@ func textState(text string, capacity, step int) []rune {
 // Exists when the context expires.
 func rollText(ctx context.Context, sd *segmentdisplay.SegmentDisplay, updateText <-chan string) {
 	colors := []cell.Color{
-		cell.ColorBlue,
+		cell.ColorNumber(33),
 		cell.ColorRed,
 		cell.ColorYellow,
-		cell.ColorBlue,
+		cell.ColorNumber(33),
 		cell.ColorGreen,
 		cell.ColorRed,
 		cell.ColorGreen,
@@ -110,7 +110,7 @@ func rollText(ctx context.Context, sd *segmentdisplay.SegmentDisplay, updateText
 }
 
 func main() {
-	t, err := termbox.New()
+	t, err := tcell.New()
 	if err != nil {
 		panic(err)
 	}
@@ -127,7 +127,7 @@ func main() {
 	go rollText(ctx, rollingSD, updateText)
 
 	input, err := textinput.New(
-		textinput.Label("New text:", cell.FgColor(cell.ColorBlue)),
+		textinput.Label("New text:", cell.FgColor(cell.ColorNumber(33))),
 		textinput.MaxWidthCells(20),
 		textinput.Border(linestyle.Light),
 		textinput.PlaceHolder("Enter any text"),